Start implementation for communication
Add a function to read a command and use it in a simple way. Need to
determine if the UI wants UCI or XBoard communication.

If only some of the source files are modified, not all the objects were
linked. This commit forces all the objects to be linked togheter.

Signed-off-by: Bogdan Popescu <bogdan.popescu0905@gmail.com>